Move BindingState interface to 'binding' package
