+++ eiffel/src/main/java/com/etiennelenhart/eiffel/state/StateViewModel.kt
@@ -8,7 +8,7 @@ import android.arch.lifecycle.ViewModel
  * A [ViewModel] exposing a single observable view state.
  *
  * To ensure an immutable state, always set a new state using the
- * [updateState] function and calling copy(...).
+ * [updateState] function and calling copy(...) on the provided current state.
  *
  * @param[T] Type of view state to expose.
  * @property[state] Observable view state.
